Remove references to Roslyn Workspaces for analyzers
Only code fixes are allowed such references.

Fixes #423
